Добавлены рекомендации ревью. Для получения даты createdAt: Date из PhotoResult.createdAt использован ISO8601DateFormatter.
