+++ ImageFeed/Helpers/Array + Extension.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+
+extension Array {
+    func withReplaced(itemAt: Index, newValue: Photo) -> [Photo] {
+        var item = ImagesListService.shared.photos
+        item.replaceSubrange(itemAt...itemAt, with: [newValue])
+        return item
+    }
+}
