+++ ImageFeed/ImagesList/ImagesListService.swift
@@ -126,11 +126,3 @@ extension ImagesListService {
         )
     }
 }
-
-extension Array {
-    func withReplaced(itemAt: Index, newValue: Photo) -> [Photo] {
-        var item = ImagesListService.shared.photos
-        item.replaceSubrange(itemAt...itemAt, with: [newValue])
-        return item
-    }
-}
