+++ ImageFeed/Models/LikeResult.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+struct LikeResult: Decodable {
+    let photo: PhotoResult?
+}
