+++ ImageFeed/Models/UrlsResult.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+struct UrlsResult: Decodable {
+    let raw: String?
+    let full: String?
+    let regular: String?
+    let small: String?
+    let thumb: String?
+}
